+++ src/core/webview/ClineProvider.ts
@@ -2471,3 +2471,4 @@ Here is the project's README to help you get started:\n\n${mcpDetails.readmeCont
 		})
 	}
 }
+